+++ archivers/zip/src/main/java/module-info.java
@@ -16,8 +16,6 @@
 
 module kala.compress.archivers.zip {
     requires transitive kala.compress.base;
-    requires static kala.compress.compressors.bzip2;
-    requires static kala.compress.compressors.deflate64;
 
     exports kala.compress.archivers.jar;
     exports kala.compress.archivers.zip;